Code to enter a person with the data already validated

+++ app/src/main/java/com/henrydev/welcomehome/data/Person.kt
@@ -14,7 +14,7 @@ data class Person(
     val firstName: String,
     @ColumnInfo(name = "last_name")
     val lastName: String,
-    val cellphone: Int,
+    val cellphone: Long,
     @ColumnInfo(name = "residential_complex")
     val residentialComplex: String,
     @ColumnInfo(name = "created_at")
